# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FST Platform 后端 API 接口文档",
        "title": "FST Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/admin/debug/pprof/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin-调试"
                ],
                "summary": "CPU profile",
                "responses": {}
            }
        },
...
        "/v1/user/avatar": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前用户的头像URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-资料"
                ],
                "summary": "更新头像",
                "parameters": [
                    {
                        "description": "头像URL {avatar: \\",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/user/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改当前登录用户的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-资料"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "密码信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_user_profile.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/user/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-资料"
                ],
                "summary": "获取个人信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_controllers_user_profile.ProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户的个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-资料"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_user_profile.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/user/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的个人设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-设置"
                ],
                "summary": "获取用户设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前用户的个人设置（语言、主题、通知偏好等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-设置"
                ],
                "summary": "更新用户设置",
                "parameters": [
                    {
                        "description": "设置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_user_profile.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_controllers_user_profile.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "app_controllers_user_profile.ProfileRealnameSummary": {
            "type": "object",
            "properties": {
                "certificateNo": {
                    "type": "string"
                },
                "certificateType": {
                    "type": "integer"
                },
                "hasVerification": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "realName": {
                    "type": "string"
                },
                "rejectReason": {
                    "type": "string"
                },
                "reviewedAt": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "submittedAt": {
                    "type": "integer"
                }
            }
        },
        "app_controllers_user_profile.ProfileResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "backGround": {
                    "type": "string"
                },
                "back_ground": {
                    "type": "string"
                },
                "birthday": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "createTime": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "groupId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "joinIp": {
                    "type": "string"
                },
                "joinTime": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "lastLoginIp": {
                    "type": "string"
                },
                "lastLoginTime": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "loginFailure": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string"
                },
                "money": {
                    "type": "number"
                },
                "motto": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "realname": {
                    "$ref": "#/definitions/app_controllers_user_profile.ProfileRealnameSummary"
                },
                "role": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "app_controllers_user_profile.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "back_ground": {
                    "type": "string"
                },
                "birthday": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "motto": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "app_controllers_user_profile.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "notify_email": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT认证令牌，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}